pyproject.toml
[project]
name = "memora"
version = "0.1.0"
description = "Memora MVP backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.6.0",
    "pypdf>=5.4.0",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "reportlab>=4.2.0",
]
